Restructure of Sensors
