+++ src/Sensors/DragSensor/index.js
@@ -1,0 +1,3 @@
+import DragSensor from './DragSensor';
+
+export default DragSensor;
