+++ src/Sensors/ForceTouchSensor/index.js
@@ -1,0 +1,3 @@
+import ForceTouchSensor from './ForceTouchSensor';
+
+export default ForceTouchSensor;
