+++ src/Sensors/MouseSensor/index.js
@@ -1,0 +1,3 @@
+import MouseSensor from './MouseSensor';
+
+export default MouseSensor;
