+++ src/Sensors/Sensor/Sensor.js
