+++ src/Sensors/Sensor/index.js
@@ -1,0 +1,3 @@
+import Sensor from './Sensor';
+
+export default Sensor;
