+++ src/Sensors/TouchSensor/index.js
@@ -1,0 +1,3 @@
+import TouchSensor from './TouchSensor';
+
+export default TouchSensor;
