+++ src/Sensors/index.js
@@ -1,0 +1,13 @@
+import Sensor from './Sensor';
+import MouseSensor from './MouseSensor';
+import TouchSensor from './TouchSensor';
+import DragSensor from './DragSensor';
+import ForceTouchSensor from './ForceTouchSensor';
+
+export {
+  Sensor,
+  MouseSensor,
+  TouchSensor,
+  DragSensor,
+  ForceTouchSensor,
+};
